Text.XML.Twiml.Internal

Datatypes à la carte

The (:+:) data type and (:<:) type class come from Swierstra's Data types à la carte.

data (f :+: g) a infixr 7

Constructors

InL (f a)

InR (g a)

class (Functor sub, Functor sup) => sub :<: sup where

Methods

inj :: sub a -> sup a

prj :: sup a -> Maybe (sub a)

Indexed

Everything in this section comes from Cirdec's excellent answer on a StackOverflow question about indexed free monads. Some names have been changed to follow the patterns established by indexed and indexed-free.

class Functor1 f where

Methods

fmap1 :: (a -> b) -> f i a -> f i b

class NFData1 f where

Methods

rnf1 :: NFData a => f i a -> ()

class Show1 f where

Methods

show1 :: Show a => f i a -> String

Applicative

class Functor1 f => IxApplicative f where

An applicative functor f indexed by a monoid (M,<>,Identity)

If you import Prelude hiding (<*>) and pure, you can redefine (<*>) and pure to use their indexed equivalents. For example,

import Prelude hiding ((<*>), pure)

pure :: IxApplicative f -> a -> f Identity a
pure = ipure

(<*>) :: IxApplicative f => f i (a -> b) -> f j a -> f (i <> j) b
(<*>) = iap

Associated Types

type Identity :: k

type i <> j :: k

Methods

ipure :: a -> f Identity a

The indexed equivalent of pure

iap :: f i (a -> b) -> f j a -> f (i <> j) b

The indexed equivalent of (<*>)

Monad

class IxApplicative m => IxMonad m where

A monad m indexed by a monoid (M,<>,Identity)

You can use do-notation with IxMonad by enabling the RebindableSyntax extension and redefining (>>=), (>>), and return. For example,

{-#LANGUAGE RebindableSyntax #-}

import Prelude hiding ((>>=), (>>), return)

(>>=) :: IxMonad m => m i a -> (a -> m j b) -> m (i <> j) b
(>>=) = ibind

(>>) :: IxMonad m => m i a -> m j b -> m (i <> j) b
a >> b = a >>= const b

return :: IxApplicative m => a -> m Identity a
return = ipure

Methods

ibind :: m i a -> (a -> m j b) -> m (i <> j) b

The indexed equivalent of (>>=)

Free

data IxFree f i a where

A free monad indexed by a monoid (M,++,[])

Constructors

IxPure :: a -> IxFree f `[]` a

IxFree :: WitnessList i => f i (IxFree f j a) -> IxFree f (i ++ j) a

iliftF :: forall f i a. (WitnessList i, Functor1 f) => f i a -> IxFree f i a

Lift an indexed functor into IxFree

type family a ++ b :: [k]

Promoted list concatenation

Equations

`[]` ++ bs = bs

(a : as) ++ bs = a : (as ++ bs)

XML

The classes here simplify working with the xml package.

class ToXML a where

Methods

toXML :: a -> [Element]

class ToElement a where

Methods

toElement :: a -> Element

class ToAttrs a where

Methods

toAttrs :: a -> [Attr]

class ToAttrValue a where

Methods

toAttrValue :: a -> String