Text.XML.Twiml.Internal.TH

Description

The code for defining TwiML verbs is highly-repetitive and follows a pattern. This module defines a little TwiML definition format and Template Haskell function for generating this code.

For example, data types Pause, PauseF, and PauseAttributes can all be generated from the following definition:

Pause
  attributes
    duration, Natural, length
  recursive
  toXMLForGADT
  toAttrsForAttributes

Documentation

data TwimlSpec

Constructors

TwimlSpec

Fields

twimlName :: String

parameters :: [Parameters]

recursive :: Bool

toXMLForGADT :: Bool

toAttrsForAttributes :: Bool

example :: String

exampleSpec :: TwimlSpec

runTwimlSpecParser :: String -> Either ParseError TwimlSpec

twimlSpecToData :: TwimlSpec -> DecsQ

Create an indexed GADT from a name. For example, given Foo, this generates (roughly)

@ data Foo data FooF i a where FooF :: a -> FooF '[Foo] a @

twimlSpecStringToData :: String -> DecsQ

s :: QuasiQuoter