Text.XML.Twiml.Types

Documentation

class ToXML a where

Methods

toXML :: a -> [Element]

class ToElement a where

Methods

toElement :: a -> Element

class ToAttrs a where

Methods

toAttrs :: a -> [Attr]

class ToAttrValue a where

Methods

toAttrValue :: a -> String

data SayAttributes

See https://www.twilio.com/docs/api/twiml/say#attributes.

Constructors

SayAttributes

Fields

_sayVoice :: Maybe Voice

_sayLoop :: Maybe Natural

data Voice

Voices supported by <Say>. See https://www.twilio.com/docs/api/twiml/say#attributes-voice.

Constructors

Man (Maybe Lang)

Woman (Maybe Lang)

Alice (Maybe LangAlice)

lang :: Voice -> Maybe (Either Lang LangAlice)

data Lang

Languages spoken by voices Man and Woman. See https://www.twilio.com/docs/api/twiml/say#attributes-manwoman.

Constructors

English

EnglishUK

Spanish

French

German

Italian

data LangAlice

Languages spoken by Alice. See https://www.twilio.com/docs/api/twiml/say#attributes-alice.

Constructors

DaDK
Danish, Denmark

DeDE
German, Germany

EnAU
English, Australia

EnCA
English, Canada

EnGB
English, UK

EnIN
English, India

EnUS
English, United States

CaES
Catalan, Spain

EsES
Spanish, Spain

EsMX
Spanish, Mexico

FiFI
Finnish, Finland

FrCA
French, Canada

FrFR
French, France

ItIT
Italian, Italy

JaJP
Japanese, Japan

KoKR
Korean, Korea

NbNO
Norwegian, Norway

NlNL
Dutch, Netherlands

PlPL
Polish-Poland

PtBR
Portuguese, Brazil

PtPT
Portuguese, Portugal

RuRU
Russian, Russia

SvSE
Swedish, Sweden

ZhCN
Chinese (Mandarin)

ZhHK
Chinese (Cantonese)

ZhTW
Chinese (Taiwanese Mandarin)

data PlayAttributes

See https://www.twilio.com/docs/api/twiml/play#attributes.

Constructors

PlayAttributes

Fields

_playLoop :: Maybe Natural

_playDigits :: Maybe [Digit]

data GatherAttributes

See https://www.twilio.com/docs/api/twiml/gather#attributes.

Constructors

GatherAttributes

Fields

_gatherAction :: Maybe URL

_gatherMethod :: Maybe Method

_gatherTimeout :: Maybe Natural

_gatherFinishOnKey :: Maybe Key

_gatherNumDigits :: Maybe Natural

data RecordAttributes

See https://www.twilio.com/docs/api/twiml/record#attributes.

Constructors

RecordAttributes

Fields

_recordAction :: Maybe URL

_recordMethod :: Maybe Method

_recordTimeout :: Maybe Natural

_recordFinishOnKey :: Maybe Key

_recordMaxLength :: Maybe Natural

_recordTranscribe :: Maybe Bool

_recordTranscribeCallback :: Maybe URL

_recordPlayBeep :: Maybe Bool

data SmsAttributes

See https://www.twilio.com/docs/api/twiml/sms#attributes.

Constructors

SmsAttributes

Fields

_smsTo :: Maybe String

_smsFrom :: Maybe String

_smsAction :: Maybe URL

_smsMethod :: Maybe Method

_smsStatusCallback :: Maybe URL

data DialAttributes

See https://www.twilio.com/docs/api/twiml/dial#attributes.

Constructors

DialAttributes

Fields

_dialAction :: Maybe URL

_dialMethod :: Maybe Method

_dialTimeout :: Maybe Natural

_dialHangupOnStar :: Maybe Bool

_dialTimeLimit :: Maybe Natural

_dialCallerId :: Maybe String

_dialRecord' :: Maybe Bool

data NumberAttributes

See https://www.twilio.com/docs/api/twiml/number#attributes.

Constructors

NumberAttributes

Fields

_numberSendDigits :: Maybe [Digit]

_numberURL :: Maybe URL

_numberMethod :: Maybe Method

data SipAttributes

See https://www.twilio.com/docs/api/twiml/sip#attributes.

Constructors

SipAttributes

Fields

_sipUsername :: Maybe String

_sipPassword :: Maybe String

_sipTransport :: Maybe Transport

_sipHeaders :: Maybe String

_sipURL :: Maybe URL

_sipMethod :: Maybe Method

data Transport

See https://www.twilio.com/docs/api/twiml/sip#transport.

Constructors

TCP

UDP

data ClientAttributes

See https://www.twilio.com/docs/api/twiml/client#attributes.

Constructors

ClientAttributes

Fields

_clientURL :: Maybe URL

_clientMethod :: Maybe Method

data ConferenceAttributes

See https://www.twilio.com/docs/api/twiml/conference#attributes.

Constructors

ConferenceAttributes

Fields

_conferenceMuted :: Maybe Bool

_conferenceBeep :: Maybe Bool

_conferenceStartOnEnter :: Maybe Bool

_conferenceEndOnExit :: Maybe Bool

_conferenceWaitURL :: Maybe URL

_conferenceWaitMethod :: Maybe Method

_conferenceMaxParticipants :: Maybe Natural

data ConferenceBeep

See https://www.twilio.com/docs/api/twiml/conference#attributes-beep.

Constructors

Yes

No

OnExit

OnEnter

data QueueAttributes

See https://www.twilio.com/docs/api/twiml/queue#attributes.

Constructors

QueueAttributes

Fields

_queueURL :: Maybe URL

_queueMethod :: Maybe Method

data DialNoun

See https://www.twilio.com/docs/api/twiml/dial#nouns.

Constructors

Number NumberAttributes String

Sip SipAttributes URL

Client ClientAttributes String

Conference ConferenceAttributes String

Queue QueueAttributes String

strToContent :: String -> Content

txtToContent :: Text -> Content

urlToContent :: URL -> Content

makeAttr' :: String -> (a -> Maybe b) -> (b -> String) -> a -> Maybe Attr

makeAttr :: ToAttrValue b => String -> (a -> Maybe b) -> a -> Maybe Attr

makeAttrs :: a -> [a -> Maybe Attr] -> [Attr]

makeElement :: Node t => String -> t -> [Attr] -> Element

data EnqueueAttributes

See https://www.twilio.com/docs/api/twiml/enqueue#attributes.

Constructors

EnqueueAttributes

Fields

_enqueueAction :: Maybe URL

_enqueueMethod :: Maybe Method

_enqueueWaitURL :: Maybe URL

_enqueueWaitURLMethod :: Maybe Method

data RedirectAttributes

See https://www.twilio.com/docs/api/twiml/redirect#attributes.

Constructors

RedirectAttributes

Fields

_redirectMethod :: Maybe Method

data RejectAttributes

See https://www.twilio.com/docs/api/twiml/reject#attributes.

Constructors

RejectAttributes

Fields

_rejectReason :: Maybe Reason

data Reason

The reason attribute takes the values "rejected" and "busy." This tells Twilio what message to play when rejecting a call. Selecting "busy" will play a busy signal to the caller, while selecting "rejected" will play a standard not-in-service response. See https://www.twilio.com/docs/api/twiml/reject#attributes-reason.

Constructors

Rejected

Busy

data PauseAttributes

See https://www.twilio.com/docs/api/twiml/pause#attributes.

Constructors

PauseAttributes

Fields

_pauseDuration :: Maybe Natural

newtype URL

Constructors

URL

Fields

getURL :: String

isHttp :: URI -> Bool

Checks whether a URI's scheme, if any, is one of "http:" or "https:".

parseURL :: String -> Maybe URL

data Method

Constructors

GET

POST

type Natural = Int

data Key

Constructors

K0
0

K1
1

K2
2

K3
3

K4
4

K5
5

K6
6

K7
7

K8
8

K9
9

KStar
*

KPound
#

data Digit

The ‘digits’ attribute lets you play DTMF tones during a call. See https://www.twilio.com/docs/api/twiml/play#attributes-digits.

Constructors

D0
0

D1
1

D2
2

D3
3

D4
4

D5
5

D6
6

D7
7

D8
8

D9
9

W
w

IxFree relies on a promoted list of type constructors, so we'll need

promoted list concatenation, (++),

a proof that (++) associates (used by iap and ibind), and

a right identity proof for [] (used by iliftF).

type family a ++ b :: [k]

Promoted list concatenation

Equations

`[]` ++ bs = bs

(a : as) ++ bs = a : (as ++ bs)

data SList i where

SList is the singleton type for promoted lists.

Constructors

Nil :: SList `[]`

Succ :: SList t -> SList (h : t)

class WitnessList xs where

Methods

witness :: SList xs

associativity :: SList xs -> Proxy ys -> Proxy zs -> (xs ++ (ys ++ zs)) :~: ((xs ++ ys) ++ zs)

A proof that (++) associates, i.e.

xs ++ (ys ++ zs) ≡ (xs ++ ys) ++ zs

rightIdentity :: SList xs -> xs :~: (xs ++ `[]`)

A proof that

xs ≡ xs ++ []

TwimlF uses ∉ in order to enforce nesting rules.

type family Elem t ts :: Bool

Elem is like a promoted elem: it allows us to check whether a type constructor t is present in a list of type constructors ts.

Equations

Elem t `[]` = False

Elem t (t : ts) = True

Elem t (u : ts) = Elem t ts

type (∉) t ts = Elem t ts ~ False

t ∉ ts is shorthand for asserting that a type constructor t is not present in a list of types constructors ts.

Every IxFunctor is a functor, so you can use fmap and (<$>) as normal.

class IxFunctor f where

An indexed functor f

Methods

imap :: (a -> b) -> f i a -> f i b

The indexed equivalent of fmap

If you import Prelude hiding (<*>) and pure, you can redefine (<*>) and pure to use their indexed equivalents. For example,

import Prelude hiding ((<*>), pure)

pure :: IxApplicative f -> a -> f Identity a
pure = ipure

(<*>) :: IxApplicative f => f i (a -> b) -> f j a -> f (i <> j) b
(<*>) = iap

class IxFunctor f => IxApplicative f where

An applicative functor f indexed by a monoid (M,<>,Identity)

Associated Types

type Identity :: k

type i <> j :: k

Methods

ipure :: a -> f Identity a

The indexed equivalent of pure

iap :: f i (a -> b) -> f j a -> f (i <> j) b

The indexed equivalent of (<*>)

You can use do-notation with IxMonad by enabling the RebindableSyntax extension and redefining (>>=), (>>), and return. For example,

{-#LANGUAGE RebindableSyntax #-}

import Prelude hiding ((>>=), (>>), return)

(>>=) :: IxMonad m => m i a -> (a -> m j b) -> m (i <> j) b
(>>=) = ibind

(>>) :: IxMonad m => m i a -> m j b -> m (i <> j) b
a >> b = a >>= const b

return :: IxApplicative m => a -> m Identity a
return = ipure

class IxApplicative m => IxMonad m where

A monad m indexed by a monoid (M,<>,Identity)

Methods

ibind :: m i a -> (a -> m j b) -> m (i <> j) b

The indexed equivalent of (>>=)

data IxFree f i a where

A free monad indexed by a monoid (M,++,[])

Constructors

IxPure :: a -> IxFree f `[]` a

IxFree :: WitnessList i => f i (IxFree f j a) -> IxFree f (i ++ j) a

iap' :: forall f i j a b. IxFunctor f => IxFree f i (a -> b) -> IxFree f j a -> IxFree f (i ++ j) b

ibind' :: forall f i j a b. IxFunctor f => IxFree f i a -> (a -> IxFree f j b) -> IxFree f (i ++ j) b

iliftF :: forall f i a. (WitnessList i, IxFunctor f) => f i a -> IxFree f i a

Lift an indexed functor into IxFree

class IxNFData f where

Methods

irnf :: NFData a => f i a -> ()

class IxShow f where

Methods

ishow :: Show a => f i a -> String

class With a where

Methods

with :: a

class HasLoop s a | s -> a where

Methods

loop :: Lens' s a

class HasVoice s a | s -> a where

Methods

voice :: Lens' s a

class HasDigits s a | s -> a where

Methods

digits :: Lens' s a

class HasAction s a | s -> a where

Methods

action :: Lens' s a

class HasFinishOnKey s a | s -> a where

Methods

finishOnKey :: Lens' s a

class HasMethod s a | s -> a where

Methods

method :: Lens' s a

class HasNumDigits s a | s -> a where

Methods

numDigits :: Lens' s a

class HasTimeout s a | s -> a where

Methods

timeout :: Lens' s a

class HasMaxLength s a | s -> a where

Methods

maxLength :: Lens' s a

class HasPlayBeep s a | s -> a where

Methods

playBeep :: Lens' s a

class HasTranscribe s a | s -> a where

Methods

transcribe :: Lens' s a

class HasTranscribeCallback s a | s -> a where

Methods

transcribeCallback :: Lens' s a

class HasFrom s a | s -> a where

Methods

from :: Lens' s a

class HasStatusCallback s a | s -> a where

Methods

statusCallback :: Lens' s a

class HasTo s a | s -> a where

Methods

to :: Lens' s a

class HasCallerId s a | s -> a where

Methods

callerId :: Lens' s a

class HasHangupOnStar s a | s -> a where

Methods

hangupOnStar :: Lens' s a

class HasRecord' s a | s -> a where

Methods

record' :: Lens' s a

class HasTimeLimit s a | s -> a where

Methods

timeLimit :: Lens' s a

class HasSendDigits s a | s -> a where

Methods

sendDigits :: Lens' s a

class HasURL s a | s -> a where

Methods

uRL :: Lens' s a

class HasHeaders s a | s -> a where

Methods

headers :: Lens' s a

class HasPassword s a | s -> a where

Methods

password :: Lens' s a

class HasTransport s a | s -> a where

Methods

transport :: Lens' s a

class HasUsername s a | s -> a where

Methods

username :: Lens' s a

class HasBeep s a | s -> a where

Methods

beep :: Lens' s a

class HasEndOnExit s a | s -> a where

Methods

endOnExit :: Lens' s a

class HasMaxParticipants s a | s -> a where

Methods

maxParticipants :: Lens' s a

class HasMuted s a | s -> a where

Methods

muted :: Lens' s a

class HasStartOnEnter s a | s -> a where

Methods

startOnEnter :: Lens' s a

class HasWaitMethod s a | s -> a where

Methods

waitMethod :: Lens' s a

class HasWaitURL s a | s -> a where

Methods

waitURL :: Lens' s a

class HasWaitURLMethod s a | s -> a where

Methods

waitURLMethod :: Lens' s a

class HasReason s a | s -> a where

Methods

reason :: Lens' s a

class HasDuration s a | s -> a where

Methods

duration :: Lens' s a