Text.XML.Twiml.Verbs.Gather

Description

The example in this file assumes

{-# LANGUAGE RebindableSyntax #-}
{-# LANGUAGE RecordWildCards #-}

import Prelude
import Text.XML.Twiml
import qualified Text.XML.Twiml.Syntax as Twiml

For more information, refer to Twilio's TwiML Reference for <Gather>.

Documentation

gather :: (IsTwimlLike f Gather, Nest i In Gather) => GatherAttributes -> TwimlLike' VoiceVerbsF i Void -> TwimlLike f Gather ()

Example:

example :: VoiceTwiml
example =
  response $ do
    gather (def & action .~ parseURL "/process_gather.php"
                 & method .~ Just GET) $ do
      say "Please enter your account number, followed by the pound sign" def
      end
    say "We didn't receive any input. Goodbye!" def
    end
  where Twiml.Syntax{..} = def

>>> show example
<?xml version="1.0" encoding="UTF-8"?>
<Response>
  <Gather action="/process_gather.php" method="GET">
    <Say>Please enter your account number, followed by the pound sign</Say>
  </Gather>
  <Say>We didn&#39;t receive any input. Goodbye!</Say>
</Response>

data Gather

data GatherF i a

data GatherAttributes

See https://www.twilio.com/docs/api/twiml/gather#attributes.