Text.XML.Twiml.Types

Documentation

data Digit

The ‘digits’ attribute lets you play DTMF tones during a call. See https://www.twilio.com/docs/api/twiml/play#attributes-digits.

Constructors

D0
0

D1
1

D2
2

D3
3

D4
4

D5
5

D6
6

D7
7

D8
8

D9
9

W
w

data Key

Constructors

K0
0

K1
1

K2
2

K3
3

K4
4

K5
5

K6
6

K7
7

K8
8

K9
9

KStar
*

KPound
#

data Method

Constructors

GET

POST

type Natural = Int

data URL

parseURL :: String -> Maybe URL

TwiMl

data VoiceTwiml

Constructors

forall i . VoiceTwiml (IxFree VoiceTwimlF i Void)

newtype VoiceTwimlF i a

Constructors

VoiceTwimlF

Fields

getVoiceTwimlF :: (SayF i :+: (PlayF i :+: (GatherF i :+: (SmsF i :+: (DialF i :+: (EnqueueF i :+: (LeaveF i :+: (HangupF i :+: (RecordF i :+: (RedirectF i :+: (RejectF i :+: (PauseF i :+: EndF i)))))))))))) a

Verbs

Say

data Say

data SayF i a where

Constructors

SayF :: String -> SayAttributes -> a -> SayF `[Say]` a

data SayAttributes

See https://www.twilio.com/docs/api/twiml/say#attributes.

Constructors

SayAttributes

Fields

_sayVoice :: Maybe Voice

_sayLoop :: Maybe Natural

data Voice

Voices supported by <Say>. See https://www.twilio.com/docs/api/twiml/say#attributes-voice.

Constructors

Man (Maybe Lang)

Woman (Maybe Lang)

Alice (Maybe LangAlice)

data Lang

Languages spoken by voices Man and Woman. See https://www.twilio.com/docs/api/twiml/say#attributes-manwoman.

Constructors

English

EnglishUK

Spanish

French

German

Italian

data LangAlice

Languages spoken by Alice. See https://www.twilio.com/docs/api/twiml/say#attributes-alice.

Constructors

DaDK
Danish, Denmark

DeDE
German, Germany

EnAU
English, Australia

EnCA
English, Canada

EnGB
English, UK

EnIN
English, India

EnUS
English, United States

CaES
Catalan, Spain

EsES
Spanish, Spain

EsMX
Spanish, Mexico

FiFI
Finnish, Finland

FrCA
French, Canada

FrFR
French, France

ItIT
Italian, Italy

JaJP
Japanese, Japan

KoKR
Korean, Korea

NbNO
Norwegian, Norway

NlNL
Dutch, Netherlands

PlPL
Polish-Poland

PtBR
Portuguese, Brazil

PtPT
Portuguese, Portugal

RuRU
Russian, Russia

SvSE
Swedish, Sweden

ZhCN
Chinese (Mandarin)

ZhHK
Chinese (Cantonese)

ZhTW
Chinese (Taiwanese Mandarin)

Play

data Play

data PlayF i a where

Constructors

PlayF :: Maybe URL -> PlayAttributes -> a -> PlayF `[Play]` a

data PlayAttributes

See https://www.twilio.com/docs/api/twiml/play#attributes.

Constructors

PlayAttributes

Fields

_playLoop :: Maybe Natural

_playDigits :: Maybe [Digit]

Gather

data Gather

type family Nest a i b

Equations

Nest i In Gather = (Record ∉ i, Gather ∉ i, Sms ∉ i, Dial ∉ i, Enqueue ∉ i, Leave ∉ i, Hangup ∉ i, Redirect ∉ i, Reject ∉ i)

data In

data GatherF i a where

Constructors

GatherF :: Nest i In Gather => GatherAttributes -> IxFree VoiceTwimlF i Void -> a -> GatherF `[Gather]` a

data GatherAttributes

See https://www.twilio.com/docs/api/twiml/gather#attributes.

Constructors

GatherAttributes

Fields

_gatherAction :: Maybe URL

_gatherMethod :: Maybe Method

_gatherTimeout :: Maybe Natural

_gatherFinishOnKey :: Maybe Key

_gatherNumDigits :: Maybe Natural

Record

data Record

data RecordF i a where

Constructors

RecordF :: RecordAttributes -> a -> RecordF `[Record]` a

data RecordAttributes

See https://www.twilio.com/docs/api/twiml/record#attributes.

Constructors

RecordAttributes

Fields

_recordAction :: Maybe URL

_recordMethod :: Maybe Method

_recordTimeout :: Maybe Natural

_recordFinishOnKey :: Maybe Key

_recordMaxLength :: Maybe Natural

_recordTranscribe :: Maybe Bool

_recordTranscribeCallback :: Maybe URL

_recordPlayBeep :: Maybe Bool

Sms

data Sms

data SmsF i a where

Constructors

SmsF :: String -> SmsAttributes -> a -> SmsF `[Sms]` a

data SmsAttributes

See https://www.twilio.com/docs/api/twiml/sms#attributes.

Constructors

SmsAttributes

Fields

_smsTo :: Maybe String

_smsFrom :: Maybe String

_smsAction :: Maybe URL

_smsMethod :: Maybe Method

_smsStatusCallback :: Maybe URL

Dial

data Dial

data DialF i a where

Constructors

DialF :: Either DialNoun String -> DialAttributes -> a -> DialF `[Dial]` a

data DialAttributes

See https://www.twilio.com/docs/api/twiml/dial#attributes.

Constructors

DialAttributes

Fields

_dialAction :: Maybe URL

_dialMethod :: Maybe Method

_dialTimeout :: Maybe Natural

_dialHangupOnStar :: Maybe Bool

_dialTimeLimit :: Maybe Natural

_dialCallerId :: Maybe String

_dialRecord' :: Maybe Bool

data DialNoun

See https://www.twilio.com/docs/api/twiml/dial#nouns.

Constructors

Number NumberAttributes String

Sip SipAttributes URL

Client ClientAttributes String

Conference ConferenceAttributes String

Queue QueueAttributes String

Number

data NumberAttributes

See https://www.twilio.com/docs/api/twiml/number#attributes.

Constructors

NumberAttributes

Fields

_numberSendDigits :: Maybe [Digit]

_numberURL :: Maybe URL

_numberMethod :: Maybe Method

Sip

data SipAttributes

See https://www.twilio.com/docs/api/twiml/sip#attributes.

Constructors

SipAttributes

Fields

_sipUsername :: Maybe String

_sipPassword :: Maybe String

_sipTransport :: Maybe Transport

_sipHeaders :: Maybe String

_sipURL :: Maybe URL

_sipMethod :: Maybe Method

data Transport

See https://www.twilio.com/docs/api/twiml/sip#transport.

Constructors

TCP

UDP

Client

data ClientAttributes

See https://www.twilio.com/docs/api/twiml/client#attributes.

Constructors

ClientAttributes

Fields

_clientURL :: Maybe URL

_clientMethod :: Maybe Method

Conference

data ConferenceAttributes

See https://www.twilio.com/docs/api/twiml/conference#attributes.

Constructors

ConferenceAttributes

Fields

_conferenceMuted :: Maybe Bool

_conferenceBeep :: Maybe Bool

_conferenceStartOnEnter :: Maybe Bool

_conferenceEndOnExit :: Maybe Bool

_conferenceWaitURL :: Maybe URL

_conferenceWaitMethod :: Maybe Method

_conferenceMaxParticipants :: Maybe Natural

data ConferenceBeep

See https://www.twilio.com/docs/api/twiml/conference#attributes-beep.

Constructors

Yes

No

OnExit

OnEnter

Queue

data QueueAttributes

See https://www.twilio.com/docs/api/twiml/queue#attributes.

Constructors

QueueAttributes

Fields

_queueURL :: Maybe URL

_queueMethod :: Maybe Method

Enqueue

data Enqueue

data EnqueueF i a where

Constructors

EnqueueF :: String -> EnqueueAttributes -> a -> EnqueueF `[Enqueue]` a

data EnqueueAttributes

See https://www.twilio.com/docs/api/twiml/enqueue#attributes.

Constructors

EnqueueAttributes

Fields

_enqueueAction :: Maybe URL

_enqueueMethod :: Maybe Method

_enqueueWaitURL :: Maybe URL

_enqueueWaitURLMethod :: Maybe Method

Leave

data Leave

data LeaveF i a where

Constructors

LeaveF :: LeaveF `[Leave]` a

Hangup

data Hangup

data HangupF i a where

Constructors

HangupF :: HangupF `[Hangup]` a

Redirect

data Redirect

data RedirectF i a where

Constructors

RedirectF :: URL -> RedirectAttributes -> RedirectF `[Redirect]` a

data RedirectAttributes

See https://www.twilio.com/docs/api/twiml/redirect#attributes.

Constructors

RedirectAttributes

Fields

_redirectMethod :: Maybe Method

Reject

data Reject

data RejectF i a where

Constructors

RejectF :: RejectAttributes -> RejectF `[Reject]` a

data RejectAttributes

See https://www.twilio.com/docs/api/twiml/reject#attributes.

Constructors

RejectAttributes

Fields

_rejectReason :: Maybe Reason

data Reason

The reason attribute takes the values "rejected" and "busy." This tells Twilio what message to play when rejecting a call. Selecting "busy" will play a busy signal to the caller, while selecting "rejected" will play a standard not-in-service response. See https://www.twilio.com/docs/api/twiml/reject#attributes-reason.

Constructors

Rejected

Busy

Pause

data Pause

data PauseF i a where

Constructors

PauseF :: PauseAttributes -> a -> PauseF `[Pause]` a

data PauseAttributes

See https://www.twilio.com/docs/api/twiml/pause#attributes.

Constructors

PauseAttributes

Fields

_pauseDuration :: Maybe Natural

End

data End

data EndF i a where

Constructors

EndF :: EndF `[End]` a

Lenses

class HasLoop s a | s -> a where

Methods

loop :: Lens' s a

class HasVoice s a | s -> a where

Methods

voice :: Lens' s a

class HasDigits s a | s -> a where

Methods

digits :: Lens' s a

class HasAction s a | s -> a where

Methods

action :: Lens' s a

class HasFinishOnKey s a | s -> a where

Methods

finishOnKey :: Lens' s a

class HasMethod s a | s -> a where

Methods

method :: Lens' s a

class HasNumDigits s a | s -> a where

Methods

numDigits :: Lens' s a

class HasTimeout s a | s -> a where

Methods

timeout :: Lens' s a

class HasMaxLength s a | s -> a where

Methods

maxLength :: Lens' s a

class HasPlayBeep s a | s -> a where

Methods

playBeep :: Lens' s a

class HasTranscribe s a | s -> a where

Methods

transcribe :: Lens' s a

class HasTranscribeCallback s a | s -> a where

Methods

transcribeCallback :: Lens' s a

class HasFrom s a | s -> a where

Methods

from :: Lens' s a

class HasStatusCallback s a | s -> a where

Methods

statusCallback :: Lens' s a

class HasTo s a | s -> a where

Methods

to :: Lens' s a

class HasCallerId s a | s -> a where

Methods

callerId :: Lens' s a

class HasHangupOnStar s a | s -> a where

Methods

hangupOnStar :: Lens' s a

class HasRecord' s a | s -> a where

Methods

record' :: Lens' s a

class HasTimeLimit s a | s -> a where

Methods

timeLimit :: Lens' s a

class HasSendDigits s a | s -> a where

Methods

sendDigits :: Lens' s a

class HasURL s a | s -> a where

Methods

uRL :: Lens' s a

class HasHeaders s a | s -> a where

Methods

headers :: Lens' s a

class HasPassword s a | s -> a where

Methods

password :: Lens' s a

class HasTransport s a | s -> a where

Methods

transport :: Lens' s a

class HasBeep s a | s -> a where

Methods

beep :: Lens' s a

class HasEndOnExit s a | s -> a where

Methods

endOnExit :: Lens' s a

class HasMaxParticipants s a | s -> a where

Methods

maxParticipants :: Lens' s a

class HasMuted s a | s -> a where

Methods

muted :: Lens' s a

class HasStartOnEnter s a | s -> a where

Methods

startOnEnter :: Lens' s a

class HasWaitMethod s a | s -> a where

Methods

waitMethod :: Lens' s a

class HasWaitURL s a | s -> a where

Methods

waitURL :: Lens' s a

class HasWaitURLMethod s a | s -> a where

Methods

waitURLMethod :: Lens' s a

class HasReason s a | s -> a where

Methods

reason :: Lens' s a

class HasDuration s a | s -> a where

Methods

duration :: Lens' s a