Text.XML.Twiml.Types

Description

Documentation

data Digit

The ‘digits’ attribute lets you play DTMF tones during a call. See https://www.twilio.com/docs/api/twiml/play#attributes-digits.

Constructors

D0
0

D1
1

D2
2

D3
3

D4
4

D5
5

D6
6

D7
7

D8
8

D9
9

W
w

data Key

Constructors

K0
0

K1
1

K2
2

K3
3

K4
4

K5
5

K6
6

K7
7

K8
8

K9
9

KStar
*

KPound
#

data Method

Constructors

GET

POST

type Natural = Int

data URL

parseURL :: String -> Maybe URL

data Voice

Voices supported by <Say>. See https://www.twilio.com/docs/api/twiml/say#attributes-voice.

Constructors

Man (Maybe Lang)

Woman (Maybe Lang)

Alice (Maybe LangAlice)

data Lang

Languages spoken by voices Man and Woman. See https://www.twilio.com/docs/api/twiml/say#attributes-manwoman.

Constructors

English

EnglishUK

Spanish

French

German

Italian

data LangAlice

Languages spoken by Alice. See https://www.twilio.com/docs/api/twiml/say#attributes-alice.

Constructors

DaDK
Danish, Denmark

DeDE
German, Germany

EnAU
English, Australia

EnCA
English, Canada

EnGB
English, UK

EnIN
English, India

EnUS
English, United States

CaES
Catalan, Spain

EsES
Spanish, Spain

EsMX
Spanish, Mexico

FiFI
Finnish, Finland

FrCA
French, Canada

FrFR
French, France

ItIT
Italian, Italy

JaJP
Japanese, Japan

KoKR
Korean, Korea

NbNO
Norwegian, Norway

NlNL
Dutch, Netherlands

PlPL
Polish-Poland

PtBR
Portuguese, Brazil

PtPT
Portuguese, Portugal

RuRU
Russian, Russia

SvSE
Swedish, Sweden

ZhCN
Chinese (Mandarin)

ZhHK
Chinese (Cantonese)

ZhTW
Chinese (Taiwanese Mandarin)

data Transport

See https://www.twilio.com/docs/api/twiml/sip#transport.

Constructors

TCP

UDP

data ConferenceBeep

See https://www.twilio.com/docs/api/twiml/conference#attributes-beep.

Constructors

Yes

No

OnExit

OnEnter

data Reason

The reason attribute takes the values "rejected" and "busy." This tells Twilio what message to play when rejecting a call. Selecting "busy" will play a busy signal to the caller, while selecting "rejected" will play a standard not-in-service response. See https://www.twilio.com/docs/api/twiml/reject#attributes-reason.

Constructors

Rejected

Busy