Text.XML.Twiml.Types

Documentation

type Natural = Int

data URL

parseURL :: String -> Maybe URL

data Method

Constructors

GET

POST

data Key

Constructors

K0
0

K1
1

K2
2

K3
3

K4
4

K5
5

K6
6

K7
7

K8
8

K9
9

KStar
*

KPound
#

data Digit

The ‘digits’ attribute lets you play DTMF tones during a call. See https://www.twilio.com/docs/api/twiml/play#attributes-digits.

Constructors

D0
0

D1
1

D2
2

D3
3

D4
4

D5
5

D6
6

D7
7

D8
8

D9
9

W
w

<Say>

data SayAttributes

See https://www.twilio.com/docs/api/twiml/say#attributes.

Constructors

SayAttributes

Fields

sayVoice :: Maybe Voice

sayLoop :: Maybe Natural

defaultSayAttributes :: SayAttributes

data Voice

Voices supported by <Say>. See https://www.twilio.com/docs/api/twiml/say#attributes-voice.

Constructors

Man (Maybe Lang)

Woman (Maybe Lang)

Alice (Maybe LangAlice)

data Lang

Languages spoken by voices Man and Woman. See https://www.twilio.com/docs/api/twiml/say#attributes-manwoman.

Constructors

English

EnglishUK

Spanish

French

German

Italian

data LangAlice

Languages spoken by Alice. See https://www.twilio.com/docs/api/twiml/say#attributes-alice.

Constructors

DaDK
Danish, Denmark

DeDE
German, Germany

EnAU
English, Australia

EnCA
English, Canada

EnGB
English, UK

EnIN
English, India

EnUS
English, United States

CaES
Catalan, Spain

EsES
Spanish, Spain

EsMX
Spanish, Mexico

FiFI
Finnish, Finland

FrCA
French, Canada

FrFR
French, France

ItIT
Italian, Italy

JaJP
Japanese, Japan

KoKR
Korean, Korea

NbNO
Norwegian, Norway

NlNL
Dutch, Netherlands

PlPL
Polish-Poland

PtBR
Portuguese, Brazil

PtPT
Portuguese, Portugal

RuRU
Russian, Russia

SvSE
Swedish, Sweden

ZhCN
Chinese (Mandarin)

ZhHK
Chinese (Cantonese)

ZhTW
Chinese (Taiwanese Mandarin)

<Play>

data PlayAttributes

See https://www.twilio.com/docs/api/twiml/play#attributes.

Constructors

PlayAttributes

Fields

playLoop :: Maybe Natural

playDigits :: Maybe [Digit]

defaultPlayAttributes :: PlayAttributes

<Gather>

data GatherAttributes

See https://www.twilio.com/docs/api/twiml/gather#attributes.

Constructors

GatherAttributes

Fields

gatherAction :: Maybe URL

gatherMethod :: Maybe Method

gatherTimeout :: Maybe Natural

gatherFinishOnKey :: Maybe Key

gatherNumDigits :: Maybe Natural

defaultGatherAttributes :: GatherAttributes

data Gather'

For some Twiml p t, the constraint (p :/~ Gather') lets us enforce TwiML nesting rules.

<Record>

data RecordAttributes

See https://www.twilio.com/docs/api/twiml/record#attributes.

Constructors

RecordAttributes

Fields

recordAction :: Maybe URL

recordMethod :: Maybe Method

recordTimeout :: Maybe Natural

recordFinishOnKey :: Maybe Key

recordMaxLength :: Maybe Natural

recordTranscribe :: Maybe Bool

recordTranscribeCallback :: Maybe URL

recordPlayBeep :: Maybe Bool

defaultRecordAttributes :: RecordAttributes

<Sms>

data SmsAttributes

See https://www.twilio.com/docs/api/twiml/sms#attributes.

Constructors

SmsAttributes

Fields

smsTo :: Maybe String

smsFrom :: Maybe String

smsAction :: Maybe URL

smsMethod :: Maybe Method

smsStatusCallback :: Maybe URL

defaultSmsAttributes :: SmsAttributes

<Dial>

data DialAttributes

See https://www.twilio.com/docs/api/twiml/dial#attributes.

Constructors

DialAttributes

Fields

dialAction :: Maybe URL

dialMethod :: Maybe Method

dialTimeout :: Maybe Natural

dialHangupOnStar :: Maybe Bool

dialTimeLimit :: Maybe Natural

dialCallerId :: Maybe String

dialRecord :: Maybe Bool

defaultDialAttributes :: DialAttributes

data DialNoun

See https://www.twilio.com/docs/api/twiml/dial#nouns.

Constructors

Number NumberAttributes String

Sip SipAttributes URL

Client ClientAttributes String

Conference ConferenceAttributes String

Queue QueueAttributes String

<Number>

data NumberAttributes

See https://www.twilio.com/docs/api/twiml/number#attributes.

Constructors

NumberAttributes

Fields

numberSendDigits :: Maybe [Digit]

numberURL :: Maybe URL

numberMethod :: Maybe Method

defaultNumberAttributes :: NumberAttributes

<Sip>

data SipAttributes

See https://www.twilio.com/docs/api/twiml/sip#attributes.

Constructors

SipAttributes

Fields

sipUsername :: Maybe String

sipPassword :: Maybe String

sipTransport :: Maybe Transport

sipHeaders :: Maybe String

sipURL :: Maybe URL

sipMethod :: Maybe Method

defaultSipAttributes :: SipAttributes

data Transport

See https://www.twilio.com/docs/api/twiml/sip#transport.

Constructors

TCP

UDP

<Client>

data ClientAttributes

See https://www.twilio.com/docs/api/twiml/client#attributes.

Constructors

ClientAttributes

Fields

clientURL :: Maybe URL

clientMethod :: Maybe Method

defaultClientAttributes :: ClientAttributes

<Conference>

data ConferenceAttributes

See https://www.twilio.com/docs/api/twiml/conference#attributes.

Constructors

ConferenceAttributes

Fields

conferenceMuted :: Maybe Bool

conferenceBeep :: Maybe Bool

conferenceStartOnEnter :: Maybe Bool

conferenceEndOnExit :: Maybe Bool

conferenceWaitURL :: Maybe URL

conferenceWaitMethod :: Maybe Method

conferenceMaxParticipants :: Maybe Natural

defaultConferenceAttributes :: ConferenceAttributes

data ConferenceBeep

See https://www.twilio.com/docs/api/twiml/conference#attributes-beep.

Constructors

Yes

No

OnExit

OnEnter

<Queue>

data QueueAttributes

See https://www.twilio.com/docs/api/twiml/queue#attributes.

Constructors

QueueAttributes

Fields

queueURL :: Maybe URL

queueMethod :: Maybe Method

defaultQueueAttributes :: QueueAttributes

<Enqueue>

data EnqueueAttributes

See https://www.twilio.com/docs/api/twiml/enqueue#attributes.

Constructors

EnqueueAttributes

Fields

enqueueAction :: Maybe URL

enqueueMethod :: Maybe Method

enqueueWaitURL :: Maybe URL

enqueueWaitURLMethod :: Maybe Method

defaultEnqueueAttributes :: EnqueueAttributes

<Redirect>

data RedirectAttributes

See https://www.twilio.com/docs/api/twiml/redirect#attributes.

Constructors

RedirectAttributes

Fields

redirectMethod :: Maybe Method

defaultRedirectAttributes :: RedirectAttributes

<Reject>

data RejectAttributes

See https://www.twilio.com/docs/api/twiml/reject#attributes.

Constructors

RejectAttributes

Fields

rejectReason :: Maybe Reason

defaultRejectAttributes :: RejectAttributes

data Reason

The reason attribute takes the values "rejected" and "busy." This tells Twilio what message to play when rejecting a call. Selecting "busy" will play a busy signal to the caller, while selecting "rejected" will play a standard not-in-service response. See https://www.twilio.com/docs/api/twiml/reject#attributes-reason.

Constructors

Rejected

Busy

<Pause>

data PauseAttributes

See https://www.twilio.com/docs/api/twiml/pause#attributes.

Constructors

PauseAttributes

Fields

pauseLength :: Maybe Natural

defaultPauseAttributes :: PauseAttributes

Lens Classes

class HasLoop t where

Methods

loop :: Lens t t (Maybe Natural) Natural

class HasAction t where

Methods

action :: Lens t t (Maybe URL) URL

class HasMethod t where

Methods

method :: Lens t t (Maybe Method) Method

class HasTimeout t where

Methods

timeout :: Lens t t (Maybe Natural) Natural

class HasFinishOnKey t where

Methods

finishOnKey :: Lens t t (Maybe Key) Key

Internal

Lens

The following section extracts a number of definitions required to get lenses, as defined in the lens package, working, without relying on the lens package itself. Rather than use the following functions, consider installing lens. See https://hackage.haskell.org/package/lens.

type Lens s t a b = forall f. Functor f => (a -> f b) -> s -> f t

type Lens' s a = Lens s s a a

lens :: (s -> a) -> (s -> b -> t) -> Lens s t a b

(^.) :: s -> Getting a s a -> a

over :: Profunctor p => Setting p s t a b -> p a b -> s -> t

to' :: (s -> a) -> IndexPreservingGetter s a

Fix & Foldable

The following definitions were extracted from the recursion-schemes package. See https://hackage.haskell.org/package/recursion-schemes.

newtype Fix f

Constructors

Fix

Fields

unFix :: f (Fix f)

class Functor (Base t) => Foldable t where

Methods

project :: t -> Base t t

cata :: (Base t a -> a) -> t -> a

type family Base t :: * -> *

Type Inequality

The following defines :/~ to mean inequality at the type level. Borrowed from HList. See http://okmij.org/ftp/Haskell/typeEQ.html.

class TypeEq x y No => x :/~ y

data Yes

data No