Text.XML.Twiml.Verbs.Gather

<Gather>

This example

module Example where

import Control.Lens
import Text.XML.Twiml

example
  = respond
  . (gather <&> timeout     .~ 10
            <&> finishOnKey .~ KStar
      $ say "Please enter your pin number and then press star."
      $ end)
  $ end

produces the following TwiML response:

<?xml version="1.0" encoding="UTF-8"?>
<Response>
  <Gather timeout="10" finishOnKey="*">
    <Say>Please enter your pin number and then press star.</Say>
  </Gather>
</Response>

data Gather p

Constructors

gather :: (Twiml Gather' n, Twiml p t, p :/~ Gather') => n -> t -> Gather p

gather' :: (Twiml Gather' n, Twiml p t, p :/~ Gather') => GatherAttributes -> n -> t -> Gather p

Attributes

data GatherAttributes

See https://www.twilio.com/docs/api/twiml/gather#attributes.

Constructors

GatherAttributes

Fields

gatherAction :: Maybe URL

gatherMethod :: Maybe Method

gatherTimeout :: Maybe Natural

gatherFinishOnKey :: Maybe Key

gatherNumDigits :: Maybe Natural

defaultGatherAttributes :: GatherAttributes

Lenses

gatherAttributes :: Lens' (Gather p) GatherAttributes

numDigits :: Lens (Gather p) (Gather p) (Maybe Natural) Natural

action :: HasAction t => Lens t t (Maybe URL) URL

method :: HasMethod t => Lens t t (Maybe Method) Method

timeout :: HasTimeout t => Lens t t (Maybe Natural) Natural

finishOnKey :: HasFinishOnKey t => Lens t t (Maybe Key) Key